requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
exceptiongroup==1.3.0
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
iniconfig==2.1.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.76
langchain-text-splitters==0.3.11
langsmith==0.4.31
lxml==6.0.2
multidict==6.6.4
numpy==2.0.2
orjson==3.11.3
packaging==25.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pypdf==6.1.0
pytest==8.4.2
python-docx==1.2.0
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
sniffio==1.3.1
tenacity==9.1.2
tokenizers==0.22.1
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
voyageai==0.3.5
yarl==1.20.1
zstandard==0.25.0
